kernel: add mprotect syscall
add curl and sdl2 ports

+++ kernel/arch/x86_64-pc/sys/syscall_handle.c
@@ -26,7 +26,7 @@ void syscall_handler(registers_t *reg) {
 
 	syscall_handle(&args);
 
-	int ret = (int)args.ret;
+	int64_t ret = (int64_t)args.ret;
 	if (ret < 0) {
 		ret = -((int)errno);
 		reg->rax = ret;
